share leaf hashes code
