+++ Cargo.toml
@@ -1,6 +1,6 @@
 [workspace]
 resolver = "2"
-members = ["host", "methods"]
+members = ["host", "methods", "shared"]
 
 # Always optimize; building and running the guest takes much longer without optimization.
 [profile.dev]
