+++ host/Cargo.toml
@@ -5,11 +5,12 @@ edition = "2021"
 
 [dependencies]
 methods = { path = "../methods" }
+shared = { path = "../shared" }
 risc0-zkvm = { version = "1.0.5" }
 tracing-subscriber = { version = "0.3", features = ["env-filter"] }
 serde = "1.0"
 rustreexo = { version = "0.3.0", features = ["with-serde"] }
-bitcoin = { version = "0.31.2", features = ["std", "rand-std", "serde"] }
+bitcoin = { version = "0.32.5", features = ["std", "rand-std", "serde"] }
 bincode = "1.3.3"
 hex = { version = "0.4.3", default-features = false, features = ["alloc"] }
 clap = { version = "4.5.16", features = ["derive"] }
