+++ host/src/main.rs
@@ -4,7 +4,7 @@ use std::fs::File;
 use methods::{METHOD_ELF, METHOD_ID};
 use risc0_zkvm::{default_prover, ExecutorEnv, Receipt};
 
-use sha2::{Digest, Sha512_256};
+use sha2::{Digest};
 use std::io::Write;
 
 use bitcoin_hashes::sha256;
@@ -17,7 +17,7 @@ use rustreexo::accumulator::stump::Stump;
 use std::str::FromStr;
 use std::time::SystemTime;
 
-use bitcoin::consensus::{deserialize, Encodable};
+use bitcoin::consensus::{deserialize};
 use bitcoin::key::Keypair;
 use bitcoin::secp256k1::{rand, Message, Secp256k1, SecretKey, Signing};
 use bitcoin::{Address, BlockHash, Network, ScriptBuf, Transaction};
@@ -25,6 +25,8 @@ use k256::schnorr;
 use k256::schnorr::signature::Verifier;
 use rustreexo::accumulator::proof::Proof;
 use serde::{Deserialize, Serialize};
+
+use shared::get_leaf_hashes;
 
 fn gen_keypair<C: Signing>(secp: &Secp256k1<C>) -> Keypair {
     let sk = SecretKey::new(&mut rand::thread_rng());
@@ -285,42 +287,4 @@ fn verify_receipt(receipt: &Receipt, s: &Stump) {
     receipt.verify(METHOD_ID).unwrap();
     println!("priv key hash: {}", sk_hash);
     println!("signed msg: {}", msg);
-}
-
-pub const UTREEXO_TAG_V1: [u8; 64] = [
-    0x5b, 0x83, 0x2d, 0xb8, 0xca, 0x26, 0xc2, 0x5b, 0xe1, 0xc5, 0x42, 0xd6, 0xcc, 0xed, 0xdd, 0xa8,
-    0xc1, 0x45, 0x61, 0x5c, 0xff, 0x5c, 0x35, 0x72, 0x7f, 0xb3, 0x46, 0x26, 0x10, 0x80, 0x7e, 0x20,
-    0xae, 0x53, 0x4d, 0xc3, 0xf6, 0x42, 0x99, 0x19, 0x99, 0x31, 0x77, 0x2e, 0x03, 0x78, 0x7d, 0x18,
-    0x15, 0x6e, 0xb3, 0x15, 0x1e, 0x0e, 0xd1, 0xb3, 0x09, 0x8b, 0xdc, 0x84, 0x45, 0x86, 0x18, 0x85,
-];
-
-fn get_leaf_hashes(
-    transaction: &Transaction,
-    vout: u32,
-    height: u32,
-    block_hash: BlockHash,
-) -> sha256::Hash {
-    let header_code = height << 1;
-
-    let mut ser_utxo = Vec::new();
-    let utxo = transaction.output.get(vout as usize).unwrap();
-    utxo.consensus_encode(&mut ser_utxo).unwrap();
-    let header_code = if transaction.is_coinbase() {
-        header_code | 1
-    } else {
-        header_code
-    };
-    let txid = transaction.txid();
-    println!("txid: {txid}, block_hash: {block_hash} vout: {vout} height: {height}");
-
-    let leaf_hash = Sha512_256::new()
-        .chain_update(UTREEXO_TAG_V1)
-        .chain_update(UTREEXO_TAG_V1)
-        .chain_update(block_hash)
-        .chain_update(transaction.txid())
-        .chain_update(vout.to_le_bytes())
-        .chain_update(header_code.to_le_bytes())
-        .chain_update(ser_utxo)
-        .finalize();
-    sha256::Hash::from_slice(leaf_hash.as_slice()).expect("parent_hash: Engines shouldn't be Err")
 }
