+++ methods/guest/Cargo.toml
@@ -6,10 +6,11 @@ edition = "2021"
 [workspace]
 
 [dependencies]
+shared = { path = "../../shared" }
 risc0-zkvm = { version = "1.0.5", default-features = false, features = ['std'] }
 rustreexo = { version = "0.3.0", features = ["with-serde"] }
 serde = "1.0"
-bitcoin = { version = "0.32.2", features = ["serde"] }
+bitcoin = { version = "0.32.5", features = ["serde"] }
 hex = { version = "0.4.3", default-features = false, features = ["alloc"] }
 sha2 = "0.10.8"
 k256 = { version = "=0.13.3", features = ["arithmetic", "serde", "expose-field", "std", "ecdsa", "pkcs8", "schnorr"], default-features = false }
