+++ shared/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "shared"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+bitcoin = { version = "0.32.5"}
+sha2 = "0.10.8"
+bitcoin_hashes = "0.14.0"
